Extract extension functions to separate file
